-convert bin-index in binning grid to long to support binning with 300 ground resolution -fix handling of lines without any bin in reprojection

+++ calvalus-binning/src/main/java/com/bc/calvalus/binning/BinManager.java
@@ -15,9 +15,9 @@ public interface BinManager {
 
     Aggregator getAggregator(int aggIndex);
 
-    SpatialBin createSpatialBin(int binIndex);
+    SpatialBin createSpatialBin(long binIndex);
 
-    TemporalBin createTemporalBin(int binIndex);
+    TemporalBin createTemporalBin(long binIndex);
 
     WritableVector createOutputVector();
